backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
pytesseract
pdf2image
Pillow
spacy
transformers
sentence-transformers
faiss-cpu
requests
openai